react-flask-app/api/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
openai
sentence_transformers
